ISSUE #868 Refactor to remove class for simplify code
