+++ bouncer/src/repo/manipulator/modelconvertor/import/odaHelper/CMakeLists.txt
@@ -7,7 +7,6 @@
 if(ODA_SUPPORT)
 	set(SOURCES
 		${SOURCES}
-		${CMAKE_CURRENT_SOURCE_DIR}/civil3d_proxy_handler.cpp
 		${CMAKE_CURRENT_SOURCE_DIR}/data_processor.cpp
 		${CMAKE_CURRENT_SOURCE_DIR}/data_processor_dgn.cpp
 		${CMAKE_CURRENT_SOURCE_DIR}/data_processor_dwg.cpp
@@ -21,7 +20,6 @@ if(ODA_SUPPORT)
 		${CMAKE_CURRENT_SOURCE_DIR}/file_processor_rvt.cpp
 		${CMAKE_CURRENT_SOURCE_DIR}/geometry_collector.cpp
 		${CMAKE_CURRENT_SOURCE_DIR}/helper_functions.cpp
-		${CMAKE_CURRENT_SOURCE_DIR}/plant3d_proxy_handler.cpp
 		${CMAKE_CURRENT_SOURCE_DIR}/repo_mesh_builder.cpp
 		${CMAKE_CURRENT_SOURCE_DIR}/repo_system_services.cpp
 		${CMAKE_CURRENT_SOURCE_DIR}/vectorise_device_rvt.cpp
@@ -30,7 +28,6 @@ if(ODA_SUPPORT)
 
 	set(HEADERS
 		${HEADERS}
-		${CMAKE_CURRENT_SOURCE_DIR}/civil3d_proxy_handler.h
 		${CMAKE_CURRENT_SOURCE_DIR}/data_processor.h
 		${CMAKE_CURRENT_SOURCE_DIR}/data_processor_dgn.h
 		${CMAKE_CURRENT_SOURCE_DIR}/data_processor_dwg.h
@@ -44,8 +41,6 @@ if(ODA_SUPPORT)
 		${CMAKE_CURRENT_SOURCE_DIR}/file_processor_rvt.h
 		${CMAKE_CURRENT_SOURCE_DIR}/geometry_collector.h
 		${CMAKE_CURRENT_SOURCE_DIR}/helper_functions.h
-		${CMAKE_CURRENT_SOURCE_DIR}/plant3d_proxy_handler.h
-		${CMAKE_CURRENT_SOURCE_DIR}/proxy_app_handler.h
 		${CMAKE_CURRENT_SOURCE_DIR}/repo_mesh_builder.h
 		${CMAKE_CURRENT_SOURCE_DIR}/repo_system_services.h
 		${CMAKE_CURRENT_SOURCE_DIR}/vectorise_device_dgn.h
