+++ bouncer/src/repo/manipulator/modelconvertor/import/odaHelper/data_processor_dwg.h
@@ -45,6 +45,20 @@ namespace repo {
 					void setMode(OdGsView::RenderMode mode);
 					~DataProcessorDwg();
 
+					/* Civil3D proxy classification. This is the only place in the DWG
+					importer that knows anything about Civil3D; removing Civil3D support
+					means deleting these three methods (and their call sites in doDraw/
+					getClassDisplayName/setEntityMetadata below). Static (no instance
+					state), so they're directly testable without constructing a
+					DataProcessorDwg. */
+					static bool isCivil3DProxyClass(const std::string& originalClass);
+
+					// TIN = Triangulated Irregular Network - Civil3D's surface-mesh
+					// representation; see AeccDbSurfaceTin/AeccDbTinSurface.
+					static bool isCivil3DSurfaceClass(const std::string& originalClass);
+
+					static bool getCivil3DDisplayName(const std::string& originalClass, std::string& outName);
+
 					// The single DwgProxyInspector instance for the file being imported,
 					// shared by every DataProcessorDwg view instance ODA creates while
 					// vectorising it. Owned by FileProcessorDwg::importModel(), alongside
@@ -78,23 +92,6 @@ namespace repo {
 						const OdGePoint3d* vertexList,
 						const OdGeVector3d* pNormal = 0) override;
 
-					void shellProc(
-						OdInt32 numVertices,
-						const OdGePoint3d* vertexList,
-						OdInt32 faceListSize,
-						const OdInt32* faceList,
-						const OdGiEdgeData* pEdgeData = 0,
-						const OdGiFaceData* pFaceData = 0,
-						const OdGiVertexData* pVertexData = 0) override;
-
-					void meshProc(
-						OdInt32 numRows,
-						OdInt32 numColumns,
-						const OdGePoint3d* vertexList,
-						const OdGiEdgeData* pEdgeData = 0,
-						const OdGiFaceData* pFaceData = 0,
-						const OdGiVertexData* pVertexData = 0) override;
-
 					void tristripProc(
 						OdInt32 numVertices,
 						const OdGePoint3d* vertexList,
@@ -108,10 +105,33 @@ namespace repo {
 						const std::string& layerId,
 						const std::string& handleMetaValue,
 						OdDbEntityPtr pEntity,
-						ProxyInfo& info,
-						ProxyGeometryCapture* capturedGeometry);
+						ProxyInfo& info);
 
 					void convertTo3DRepoColor(OdCmEntityColor& color, repo::lib::repo_color3d_t& out);
+
+					/* Derives and draws the 3 (deduped) edges of one triangle for the
+					current surface-type entity, as a wireframe overlay. Returns false,
+					doing nothing, for a degenerate triangle (two coincident points).
+					Only called when drawingSurfaceEdges is true - see doDraw. */
+					bool addSurfaceTriangle(
+						const repo::lib::RepoVector3D64& p0,
+						const repo::lib::RepoVector3D64& p1,
+						const repo::lib::RepoVector3D64& p2);
+
+					void addSurfaceEdgeIfNeeded(
+						const repo::lib::RepoVector3D64& p0,
+						const repo::lib::RepoVector3D64& p1);
+
+					// Civil3D extension-dictionary metadata contribution - see
+					// setEntityMetadata, called right after proxy->getProxyEntityMetadata()
+					// returns (info.extensionDictionary is already populated by then).
+					static void addCivil3DDictionaryMetadata(
+						const ProxyInfo& info,
+						std::unordered_map<std::string, repo::lib::RepoVariant>& metadata);
+
+					// "Civil3D (<class>)" / "CustomApp (<class>)" / XData-based fallback,
+					// for the "Proxy::Application" metadata field.
+					static std::string formatProxyApplicationString(const ProxyInfo& info);
 
 					class Layer
 					{
@@ -134,6 +154,20 @@ namespace repo {
 						}
 					};
 
+					// Runs the actual draw call for one entity/drawable within doDraw(),
+					// with Civil3D proxy/TIN-surface awareness: dispatches to stored
+					// proxy graphics when isProxy is set, and scopes the surface-edge
+					// overlay state (see addSurfaceTriangle) to this one draw. Extracted
+					// from doDraw() because this whole chunk - proxy dispatch and
+					// surface-edge state - has no equivalent for non-proxy entities.
+					bool drawProxyAwareGeometry(
+						OdUInt32 i,
+						const OdGiDrawable* pDrawable,
+						OdDbEntityPtr pEntity,
+						bool isProxy,
+						const ProxyInfo& info,
+						GeometryCollector::Context* ctx);
+
 					// Some properties to be held between invocations of doDraw()
 					class Context
 					{
@@ -148,6 +182,16 @@ namespace repo {
 
 					// Not owned; see setProxy().
 					DwgProxyInspector* proxy = nullptr;
+
+					/* State for the current entity's surface-edge rendering (see
+					addSurfaceTriangle/addSurfaceEdgeIfNeeded). Both are fully saved
+					and restored around each doDraw() call - not just reset on
+					transition - so that sibling or nested surface-type entities never
+					leak state into one another; a file with several TIN surfaces must
+					produce one distinct mesh (plus one distinct edge mesh) per surface
+					entity. */
+					bool drawingSurfaceEdges = false;
+					std::unordered_set<std::string> currentSurfaceEdgeKeys;
 
 					std::string getClassDisplayName(OdDbEntityPtr entity, const ProxyInfo& info);
 				};
