+++ test/src/unit/repo/manipulator/modelconvertor/import/odaImport/ut_repo_model_import_oda.cpp
@@ -195,41 +195,28 @@ TEST(ODAModelImport, ColouredBoxesDWG)
 
 TEST(ODAModelImport, Civil3DTinSurfaceDWG)
 {
-	// This fixture contains a Civil3D TIN surface, which DWG proxy-imports
-	// as a AeccDbSurfaceTin/AeccDbTinSurface proxy entity with no native
-	// geometry - Civil3DProxyHandler::TinCapture reconstructs the mesh by
-	// capturing triangles/edges from the entity's stored proxy graphics as
-	// they're replayed (see DataProcessorDwg::processTriangleOut et al.).
-
-	// NOTE: filename is a placeholder pending the real fixture being added
-	// to the REPO_MODEL_PATH test data set.
+	// This fixture contains a Civil3D TIN surface, which DWG proxy-imports as
+	// an AeccDbSurfaceTin/AeccDbTinSurface proxy entity with no native
+	// geometry - Civil3DProxyHandler::isSurfaceClass() flags it as a surface,
+	// and DataProcessorDwg::addSurfaceTriangle draws the replayed triangles
+	// as a single shaded mesh plus a single wireframe edge mesh (see
+	// DataProcessorDwg::processTriangleOut et al.), rather than one mesh per
+	// triangle.
 	auto scene = ODAModelImportUtils::ModelImportManagerImport("Civil3DTinSurfaceDWG", getDataPath("Proxy_Civil3d_TinSurface.dwg"));
 	SceneUtils utils(scene);
 
-	// DataProcessorDwg::doDraw's tinSurfaceProxy branch attaches the captured
-	// faces and metadata to the DWG *layer* node (parentLayer.id) rather than
-	// creating a distinct node for the surface entity itself - so this looks
-	// up that layer node via the metadata TinCapture::addComputedMetadata
-	// sets on it, rather than by an entity display name.
+	// The surface's meshes and metadata are attached to the entity's own
+	// layer node (entityLayer.id), found here via the metadata
+	// DataProcessorDwg::setEntityMetadata sets on it.
 	auto surfaceLayerNodes = utils.findNodesByMetadata("Entity Handle::Value", "[43A2]");
 	ASSERT_THAT(surfaceLayerNodes.size(), Gt(0));
 	auto surfaceLayer = surfaceLayerNodes[0];
 
-	// TinCapture::applyFaceLayers creates one "Face" child layer per
-	// captured triangle, each carrying its own mesh.
-	auto faces = surfaceLayer.getChildren({ repo::core::model::NodeType::TRANSFORMATION });
-	ASSERT_THAT(faces.size(), Gt(0));
-	for (auto& face : faces) {
-		EXPECT_THAT(face.name(), Eq("Face"));
-		EXPECT_THAT(face.hasGeometry(), IsTrue());
-	}
-
-	// TinCapture::addComputedMetadata computes these from the captured triangles.
-	auto metadata = surfaceLayer.getMetadata();
-
-	// Elevation range is data-dependent, so only presence is checked for now.
-	EXPECT_THAT(metadata.count("Data::Minimum Elevation"), Eq(1u));
-	EXPECT_THAT(metadata.count("Data::Maximum Elevation"), Eq(1u));
+	// addSurfaceTriangle forces every triangle under one fixed material (the
+	// shaded mesh) and derives a deduped wireframe overlay (the edge mesh),
+	// both attached directly to the surface's own node - no per-face child
+	// layers.
+	EXPECT_THAT(surfaceLayer.hasGeometry(), IsTrue());
 }
 
 MATCHER_P(Paths, matcher, "") {
